fix(comment): dedupe overlay-skipping picker, dismiss empty draft on click-away
- Hoist pickContentElementAtPoint into dom-element-utils; comment-hover-overlay
  and page-content both call the shared helper.
- Click-away on an empty composer clears the draft instead of committing a new
  anchor; read draft state via ref so the pointer effect doesn't rebind on
  every keystroke.
